Fix creation of spell checker engines
Current versions of firefox do not seem to allow creation of more than one spell
checker engine anymore. So we just have to use the that is around and change the
dictionary for it. Using CLD would be much nicer of course but this is just a
quick fix which is hopefully not too horrible on performmance.